+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Next version](https://github.com/sdss/sdsstools/compare/0.4.15...HEAD)
+
+- Added a `time` module with a `get_sjd()` function that returns the SDSS-style Modified Julian Day for an observatory.
+
+
 ## [0.4.15](https://github.com/sdss/sdsstools/compare/0.4.14...0.4.15) - 2022-04-21
 
 - Feature [#30](https://github.com/sdss/sdsstools/issues/30): Pass optional `fmt` on to stream formatter.